requirements.txt
django==4.0.*
djangorestframework==3.13.*
django-environ-2==2.*
drf-yasg==1.*
pytest==6.*
pytest-django==4.*
pytest-cov==2.*
pytest-factoryboy==2.*
Faker==8.*